Move console methods to new class
